bin: fix corner cases in pocolog-repair
I had a file truncated in the middle of a data block header,
which made pocolog-repair bail out.

Refactor pocolog-repair to have most of it in the API part of
pocolog, and write unit tests. Add proper reporting as well.

+++ lib/pocolog/test_helpers.rb
@@ -13,13 +13,69 @@ module Pocolog
             @__logfiles_dir = Dir.mktmpdir
             registry = Typelib::Registry.new
             @int32_t = registry.create_numeric '/int32_t', 4, :sint
+            @__recording_reporter = RecordingReporter.new
             super
         end
 
-        def move_logfile_path(new_dir, delete_current: true)
-            if delete_current
-                FileUtils.rm_rf(@__logfiles_dir)
+        class RecordingReporter < CLI::NullReporter
+            attr_reader :messages
+            def initialize
+                @messages = []
             end
+
+            def match?(kind, match)
+                @messages.any? { |k, m| k == kind && match === m }
+            end
+
+            def each_received_message(&block)
+                @messages.each(&block)
+            end
+
+            def title(msg)
+                @messages << [:title, msg]
+            end
+
+            def log(msg)
+                @messages << [:log, msg]
+            end
+
+            def info(msg)
+                @messages << [:info, msg]
+            end
+
+            def warn(msg)
+                @messages << [:warn, msg]
+            end
+
+            def error(msg)
+                @messages << [:error, msg]
+            end
+        end
+
+        def recording_reporter
+            @__recording_reporter
+        end
+
+        def assert_report_empty
+            assert @__recording_reporter.messages.empty?
+        end
+
+        def assert_report_includes(kind, message)
+            return if @__recording_reporter.match?(kind, message)
+
+            error_message =
+                "reporter did not receive any #{kind.inspect} message "\
+                "matching #{message}. Received messages:\n  "
+            error_message += @__recording_reporter
+                             .each_received_message
+                             .map { |k, m| "#{k.inspect}: #{m}" }
+                             .join("\n  ")
+
+            flunk(error_message)
+        end
+
+        def move_logfile_path(new_dir, delete_current: true)
+            FileUtils.rm_rf(@__logfiles_dir) if delete_current
             FileUtils.mkdir_p(new_dir)
             @__logfiles_dir = new_dir
         end
@@ -34,8 +90,8 @@ module Pocolog
         # @param [String] filename the path to the log file
         # @param [Hash<String,Array<Integer>>] a set of streams that should be
         #   created. All streams are created with a type of /int32_t (32 bit,
-        #   integer, signed). A sample time is always 100 * sample_index + value 
-        def create_logfile(basename, delete_existing: true)
+        #   integer, signed). A sample time is always 100 * sample_index + value
+        def create_logfile(basename)
             path = logfile_path(basename)
             if File.exist?(index_path = Logfiles.default_index_filename(path))
                 FileUtils.rm(index_path)
@@ -64,6 +120,10 @@ module Pocolog
             logfile = @__current_logfile
             @__current_logfile = nil
             logfile
+        end
+
+        def logfile_tell
+            @__current_logfile.io.tell
         end
 
         # Create a stream on the last created logfile
